Changes: - move multichecker to /pkg dir
